requirements.txt
flask
boto3
google-genai
PyPDF2
python-pptx
python-docx
gunicorn
